Plumbs `scrollBehavior` into `SelectableText` so that the scrollbar may be hidden (#158887)
## Description

This adds a `scrollBehavior` attribute to `SelectableText` so that the
scrolling can be controlled more directly.

I added this specifically because it's not possible to turn off the
scrollbar on a selectable text, even if you set the scroll physics to be
`NeverScrollableScrollPhysics`. We had a UI where we needed to have a
clipped, multi-line selectable text field, but have it not be scrollable
unless it was expanded in size, and it wasn't possible to hide the
scrollbar, but still wanted it to be selectable.

## Tests
- Added a test that makes sure that the scroll behavior makes it down to
the `Scrollable` in the `EditableText`.